Time: 19 ms (65.75%), Space: 6 MB (31.51%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if(x < 0) return false;
-        long long k = 0;
+        long k = 0;
         int num = x;
         while(x > 0){
             int lsb = x % 10;
